+++ pcs/usage.py
@@ -1568,7 +1568,7 @@ Commands:
         resource name regular expression regexp%<resource_pattern>.
 
     location <resource> rule [id=<rule id>] [resource-discovery=<option>]
-             [role=master|slave] [constraint-id=<id>]
+             [role=Promoted|Unpromoted] [constraint-id=<id>]
              [score=<score> | score-attribute=<attribute>] <expression>
         Creates a location constraint with a rule on the specified resource
         where expression looks like one of the following:
@@ -1651,16 +1651,16 @@ Commands:
         mean the resources should not be run on the same node.  Specifying
         'INFINITY' (or '-INFINITY') for the score forces <source resource> to
         run (or not run) with <target resource> (score defaults to "INFINITY").
-        A role can be: 'Master', 'Slave', 'Started', 'Stopped' (if no role is
-        specified, it defaults to 'Started').
+        A role can be: 'Promoted', 'Unpromoted', 'Started', 'Stopped' (if no
+        role is specified, it defaults to 'Started').
 
     colocation set <resource1> [resourceN]... [options]
                [set <resourceX> ... [options]]
                [setoptions [constraint_options]]
         Create a colocation constraint with a resource set.
         Available options are sequential=true/false and
-        role=Stopped/Started/Master/Slave. Available constraint_options are id
-        and either of: score, score-attribute, score-attribute-mangle.
+        role=Stopped/Started/Promoted/Unpromoted. Available constraint_options
+        are id and either of: score, score-attribute, score-attribute-mangle.
 
     colocation delete <source resource id> <target resource id>
         Remove colocation constraints with specified resources.
@@ -1676,13 +1676,13 @@ Commands:
                [id=<constraint-id>]
         Create a ticket constraint for <resource id>.
         Available option is loss-policy=fence/stop/freeze/demote.
-        A role can be master, slave, started or stopped.
+        A role can be Promoted, Unpromoted, Started or Stopped.
 
     ticket set <resource1> [<resourceN>]... [<options>]
                [set <resourceX> ... [<options>]]
                setoptions <constraint_options>
         Create a ticket constraint with a resource set.
-        Available options are role=Stopped/Started/Master/Slave. Required
+        Available options are role=Stopped/Started/Promoted/Unpromoted. Required
         constraint option is ticket=<ticket>. Optional constraint options are
         id=<constraint-id> and loss-policy=fence/stop/freeze/demote.
 
@@ -1701,7 +1701,7 @@ Commands:
     ref <resource>...
         List constraints referencing specified resource.
 
-    rule add <constraint id> [id=<rule id>] [role=master|slave]
+    rule add <constraint id> [id=<rule id>] [role=Promoted|Unpromoted]
              [score=<score>|score-attribute=<attribute>] <expression>
         Add a rule to a location constraint specified by 'constraint id' where
         the expression looks like one of the following: